rebase + a bit more changes
remove useless feature
