+++ esp-hal/src/i2s/parallel.rs
@@ -39,7 +39,7 @@
 //!
 //! ```rust, no_run
 #![doc = crate::before_snippet!()]
-//! # use esp_hal::dma::{Dma, DmaPriority, DmaTxBuf};
+//! # use esp_hal::dma::{Dma, DmaTxBuf};
 //! # use esp_hal::dma_buffers;
 //! # use esp_hal::delay::Delay;
 //! # use esp_hal::i2s::parallel::{I2sParallel, TxEightBits};
@@ -67,11 +67,11 @@
 //! let (_, _, tx_buffer, tx_descriptors) = dma_buffers!(0, BUFFER_SIZE);
 //! let mut parallel = I2sParallel::new(
 //!     i2s,
-//!     dma_channel.configure(false, DmaPriority::Priority0),
+//!     dma_channel,
 //!     1.MHz(),
 //!     pins,
 //!     clock,
-//! );
+//! ).into_async();
 //!
 //! for (i, data) in tx_buffer.chunks_mut(4).enumerate() {
 //!     let offset = i * 4;
