--- qa-test/src/bin/i2s_sound.rs
@@ -1,110 +1,0 @@
-//! This shows how to transmit data continuously via I2S.
-//!
-//! Without an additional I2S sink device you can inspect the MCLK, BCLK, WS
-//!  and DOUT with a logic analyzer.
-//!
-//! You can also connect e.g. a PCM510x to hear an annoying loud sine tone (full
-//! scale), so turn down the volume before running this example.
-//!
-//! The following wiring is assumed:
-//! - BCLK => GPIO2
-//! - WS   => GPIO4
-//! - DOUT => GPIO5
-//!
-//! PCM510x:
-//! | Pin   | Connected to    |
-//! |-------|-----------------|
-//! | BCK   | GPIO0           |
-//! | DIN   | GPIO2           |
-//! | LRCK  | GPIO1           |
-//! | SCK   | Gnd             |
-//! | GND   | Gnd             |
-//! | VIN   | +3V3            |
-//! | FLT   | Gnd             |
-//! | FMT   | Gnd             |
-//! | DEMP  | Gnd             |
-//! | XSMT  | +3V3            |
-
-//% CHIPS: esp32 esp32c3 esp32c6 esp32h2 esp32s2 esp32s3
-
-#![no_std]
-#![no_main]
-
-use esp_backtrace as _;
-use esp_hal::{
-    dma::Dma,
-    dma_buffers,
-    i2s::master::{DataFormat, I2s, Standard},
-    prelude::*,
-};
-
-const SINE: [i16; 64] = [
-    0, 3211, 6392, 9511, 12539, 15446, 18204, 20787, 23169, 25329, 27244, 28897, 30272, 31356,
-    32137, 32609, 32767, 32609, 32137, 31356, 30272, 28897, 27244, 25329, 23169, 20787, 18204,
-    15446, 12539, 9511, 6392, 3211, 0, -3211, -6392, -9511, -12539, -15446, -18204, -20787, -23169,
-    -25329, -27244, -28897, -30272, -31356, -32137, -32609, -32767, -32609, -32137, -31356, -30272,
-    -28897, -27244, -25329, -23169, -20787, -18204, -15446, -12539, -9511, -6392, -3211,
-];
-
-#[entry]
-fn main() -> ! {
-    let peripherals = esp_hal::init(esp_hal::Config::default());
-
-    let dma = Dma::new(peripherals.DMA);
-    #[cfg(any(feature = "esp32", feature = "esp32s2"))]
-    let dma_channel = dma.i2s0channel;
-    #[cfg(not(any(feature = "esp32", feature = "esp32s2")))]
-    let dma_channel = dma.channel0;
-
-    let (_, rx_descriptors, tx_buffer, tx_descriptors) = dma_buffers!(0, 32000);
-
-    let i2s = I2s::new(
-        peripherals.I2S0,
-        Standard::Philips,
-        DataFormat::Data16Channel16,
-        44100.Hz(),
-        dma_channel,
-        rx_descriptors,
-        tx_descriptors,
-    );
-
-    let mut i2s_tx = i2s
-        .i2s_tx
-        .with_bclk(peripherals.GPIO2)
-        .with_ws(peripherals.GPIO4)
-        .with_dout(peripherals.GPIO5)
-        .build();
-
-    let data =
-        unsafe { core::slice::from_raw_parts(&SINE as *const _ as *const u8, SINE.len() * 2) };
-
-    let mut idx = 0;
-    for i in 0..usize::min(data.len(), tx_buffer.len()) {
-        tx_buffer[i] = data[idx];
-
-        idx += 1;
-
-        if idx >= data.len() {
-            idx = 0;
-        }
-    }
-
-    let mut filler = [0u8; 10000];
-    let mut transfer = i2s_tx.write_dma_circular(&tx_buffer).unwrap();
-
-    loop {
-        let avail = transfer.available().unwrap();
-        if avail > 0 {
-            let avail = usize::min(10000, avail);
-            for bidx in 0..avail {
-                filler[bidx] = data[idx];
-                idx += 1;
-
-                if idx >= data.len() {
-                    idx = 0;
-                }
-            }
-            transfer.push(&filler[0..avail]).unwrap();
-        }
-    }
-}
